fix: symbol table order #1
fixing the issue where the declaration of the
parameters where included in the mother table,
and not in the function table. currently, I am not
registering the arguments anywhere but the func
symbol table, what leads me to the problem where
I can not check the arguments on a random call to
it.

So I must save this info in the structure with
the function information, to be able to check it
(for checkin the argument types) later [besides
saving it to symbtable as well, a design choice]